renamed package for custom swing elements, wrote javadocs for baccarat interface adapters
