+++ src/view/baccarat/BaccaratStartView.java
@@ -5,7 +5,7 @@ import interface_adapter.baccarat.BaccaratStartState;
 import interface_adapter.baccarat.BaccaratStartViewModel;
 import interface_adapter.game_menu.exit.ExitController;
 import tools.GridBagUtils;
-import view.custom_elements.*;
+import view.custom_swing_elements.*;
 
 import javax.swing.*;
 import javax.swing.event.DocumentEvent;
