+++ src/view/blackjack/BlackJackIngameView.java
@@ -3,8 +3,8 @@ package view.blackjack;
 import interface_adapter.blackjack.blackjack_logic.*;
 import interface_adapter.game_menu.exit.ExitController;
 import tools.GridBagUtils;
-import view.custom_elements.BlackJackBackgroundPanel;
-import view.custom_elements.GreenCustomButton;
+import view.custom_swing_elements.BlackJackBackgroundPanel;
+import view.custom_swing_elements.GreenCustomButton;
 
 import javax.swing.*;
 import java.awt.*;
